Fix tooltip theme contrast
