+++ application/app/topTabsChromeTheme.ts
@@ -1,4 +1,5 @@
 import type { TerminalTheme } from '../../types';
+import { resolveReadableForegroundForHsl } from '../../domain/colorContrast';
 
 function hexToHslToken(hex: string): string {
   const normalized = hex.startsWith('#') ? hex : `#${hex}`;
@@ -61,7 +62,9 @@ const TOP_TABS_THEME_PROPERTIES = [
   '--background',
   '--foreground',
   '--accent',
+  '--accent-foreground',
   '--primary',
+  '--primary-foreground',
   '--secondary',
   '--border',
   '--muted-foreground',
@@ -84,6 +87,7 @@ export function applyTopTabsChromeThemeVars(theme: TerminalTheme): void {
   const bg = hexToHslToken(theme.colors.background);
   const fg = hexToHslToken(theme.colors.foreground);
   const accent = hexToHslToken(theme.colors.cursor);
+  const accentForeground = resolveReadableForegroundForHsl(accent);
   const isDark = theme.type === 'dark';
   const secondary = adjustLightnessToken(bg, isDark ? 6 : -5);
   const border = adjustLightnessToken(bg, isDark ? 12 : -10);
@@ -92,7 +96,9 @@ export function applyTopTabsChromeThemeVars(theme: TerminalTheme): void {
   setStylePropertyIfChanged(tabsRoot, '--background', bg);
   setStylePropertyIfChanged(tabsRoot, '--foreground', fg);
   setStylePropertyIfChanged(tabsRoot, '--accent', accent);
+  setStylePropertyIfChanged(tabsRoot, '--accent-foreground', accentForeground);
   setStylePropertyIfChanged(tabsRoot, '--primary', accent);
+  setStylePropertyIfChanged(tabsRoot, '--primary-foreground', accentForeground);
   setStylePropertyIfChanged(tabsRoot, '--secondary', secondary);
   setStylePropertyIfChanged(tabsRoot, '--border', border);
   setStylePropertyIfChanged(tabsRoot, '--muted-foreground', mutedFg);
